separated Logger class out from IfcUtil file. Added SetProduct function to Logger class and plugged that into IfcGeomIterator.h
